pyproject.toml
[project]
name = "cves-analytics"
version = "0.2.0"
description = "CVE data analytics with simulation scenario generation and vulnerability analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "codeium-jupyter>=2.1.2",
    "cve-bin-tool>=3.4",
    "docker>=7.1.0",
    "faker>=36.1.1",
    "flask>=3.1.0",
    "gensim>=4.3.3",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "jupyterlab>=4.3.4",
    "keras>=3.7.0",
    "networkx>=3.4.2",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pgmpy>=0.1.26",
    "pip>=6.0.0",
    "plotly",
    "polars>=1.33.0",
    "pyarrow>=18.1.0",
    "pydantic-settings>=2.12.0",
    "pyexploitdb>=0.2.61",
    "pygrype>=0.4.0",
    "pyldavis>=3.4.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.8.4",
    "scikit-learn>=1.6.0",
    "seaborn>=0.13.2",
    "reportlab>=4.0.0",
    "spacy>=3.8.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.8.4",
    "uv>=0.5.13",
]

[tool.ruff]
# Same behavior as Black for line length
line-length = 88

# Enable Flake8 (E, F), isort (I), pyupgrade (UP),
# and flake8-bugbear (B) rules
lint.select = ["E", "F", "I", "UP", "B"]

# Exclude directories you don't want to analyze
lint.exclude = [".venv", "logs", "docs", "output", "uploads", "data", "notebooks"]

[tool.ruff.lint.isort]
# Help Ruff differentiate your imports from third-party ones
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
# E501 line length - many files have complex expressions that are hard to break
"src/**/*.py" = ["E501"]
"tests/**/*.py" = ["E501"]
